requirements.txt
torch>=2.1.0
numpy
torchvision
kornia>=0.7.0
Pillow
timm==0.9.8
tqdm
onedrivedownloader
ftfy
regex
pyyaml
snntorch